REFACTOR base class abstraction
NEW \Dynamic\CsvUtility\Utility\SilverStripeCsvUtility
 - Wrapper class handling SilverStripe specific data logic

NEW \Dynamic\CsvUtility\CsvUtilTraits\CsvUtilityTrait
- A trait allowing for DataExtension type hooks for a pre-processing method `preProcess($data)`

closes #2

closes #
