+++ README.md
@@ -12,56 +12,4 @@ See the [composer.json](composer.json) file.
 
 ## Example usage
 
-```php
-
-<?php
-
-class MyDataObject extends DataObjet implements CsvUtilityInterface
-{
-    
-    private static $db = [
-        'FirstName' => 'Varchar',
-        'LastName' => 'Varchar',
-    ];
-    
-    public function getExportFields() {
-        return [
-            'FirstName' => 'First Name',
-            'LastName' => 'Last Name',
-        ];
-    }
-    
-}
-
-class MyController extends Page_Controller
-{
-    
-    public function getcsv(SS_HTTPRequest $request)
-    {
-        
-        $utility = new dynamic\CsvUtility\CsvUtility(
-            MyDataObject::get()
-                ->filter([
-                    'Foo' => 'a',
-                    'Bar' => 'b',
-                ]), $request
-        );
-        
-        $utility->setAllowedReportTypes([
-            'MyRequestVarToCheckFor' => 'MyDataObject',
-        ]);
-        
-        $csv = $utility->getFileContents();
-        
-        if ($csv) {
-            return SS_HTTPRequest::send_file($csv, $fileName = 'MyFileName.csv', 'Content-Type: text/csv');
-        }
-        
-        return $this->httpError(404);
-        
-    }
-
-}
-
-
-```
+See [usage documentation](docs/en/usage/index.md) for specific usage types
