+++ src/traits/CsvUtilityTrait.php
@@ -1,0 +1,21 @@
+<?php
+
+namespace Dynamic\CsvUtility\CsvUtilTraits;
+
+/**
+ * Class CsvUtilityTrait
+ * @package Dynamic\CsvUtility\CsvUtilTraits
+ */
+trait CsvUtilityTrait
+{
+
+    /**
+     * @param $data
+     * @return mixed
+     */
+    public function preProcessData($data)
+    {
+        return $data;
+    }
+
+}
